M analysis/server/j2se/com/sun/sgs/analysis/task/BasicScheduleTasks.java
- Remove latency number -- we can't compute this here since we don't
  know the scheduling and completion times of each task.

M src/server/j2se/com/sun/sgs/impl/profile/listener/ProfileSummaryListener.java

- Remove a redundant cast.

+++ src/server/j2se/com/sun/sgs/impl/profile/listener/ProfileSummaryListener.java
@@ -239,7 +239,7 @@ public class ProfileSummaryListener implements ProfileListener {
 			      ((lifetimeSuccessful*1000) /
 			       (double)lifetimeWindowTime),
 			      ((lifetimeRunTime + lifetimeLagTime) /
-			       (double)lifetimeSuccessful)
+			       lifetimeSuccessful)
 		);
 
  	    maxRunTime = 0;
